Refactor import statements and organize enum declarations
